插件服务代码重构
